% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_wordnet}
\alias{download_wordnet}
\alias{delete_wordnet}
\title{Download And Delete WordNet SQL Database}
\usage{
download_wordnet(debug = FALSE)

delete_wordnet()
}
\arguments{
\item{debug}{a flag for debugging. You should keep it FALSE. (Try at your own risk!)}
}
\value{
TRUE if the database is found. FALSE if there is no database and it is not running in an interactive R session.
}
\description{
To download the WordNet 3.1 data as a sqlite database. It runs only in an interactive R session. The size of the database is around 500MB. Please make sure you agree with the WordNet License.
}
