% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehr.R
\name{get_lemmas}
\alias{get_lemmas}
\title{Search For Lemmas in WordNet}
\usage{
get_lemmas(x = c("very", "nice"), pos = c("n", "v", "a", "s", "r"), sensenum)
}
\arguments{
\item{x}{character, one or more lemmas to be searched}

\item{pos}{character, a vector of part-of-speech labels: "n": Noun, "v": Verb, "a": Adjective, "s": Adjective satellite, "r": Adverb}

\item{sensenum}{integer, if supplied, only those sensenum are selected.}
}
\value{
a data frame containing search result
}
\description{
Search for lemma(s) in WordNet.
}
