% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehr.R
\name{get_synsetids}
\alias{get_synsetids}
\alias{get_synonyms}
\title{Search for Synset IDs in WordNet}
\usage{
get_synsetids(x = c("301590922", "108957024"))

get_synonyms(x)
}
\arguments{
\item{x}{character, one or more Synset IDs to be searched, or a data.frame result from another \code{get_} function}
}
\value{
a data frame containing search result
}
\description{
Search for Synset ID(s) in WordNet
}
\examples{
\donttest{
if (interactive()) {
  get_synsetids("106618544")
}
}
}
