% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehr.R
\name{search_synsetid}
\alias{search_synsetid}
\title{Search for Synset ID in WordNet}
\usage{
search_synsetid(synsetid = c("301590922", "108957024"))
}
\arguments{
\item{synsetid}{character, one or more Synset IDs to be searched}
}
\value{
a data frame containing search result
}
\description{
Search for Synset ID(s) in WordNet
}
